Move site metadata
