+++ src/pages/about.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
+import PageInfo from "../components/site-metadata/PageInfo";
 import AboutPage from "../page-about/AboutPage";
-import PageInfo from "../site-metadata/PageInfo";
 
 // eslint-disable-next-line react/display-name
 export default () => (
