+++ src/pages/facility.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
 import AuthWall from "../auth/AuthWall";
+import PageInfo from "../components/site-metadata/PageInfo";
 import FacilityPage from "../page-multi-facility/FacilityPage";
-import PageInfo from "../site-metadata/PageInfo";
 
 // eslint-disable-next-line react/display-name
 export default () => (
