+++ src/pages/get-involved.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
+import PageInfo from "../components/site-metadata/PageInfo";
 import GetInvolvedPage from "../page-get-involved/GetInvolvedPage";
-import PageInfo from "../site-metadata/PageInfo";
 
 // eslint-disable-next-line react/display-name
 export default () => (
