+++ src/pages/verify.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
+import PageInfo from "../components/site-metadata/PageInfo";
 import VerificationNeeded from "../page-verification-needed/VerificationNeeded";
-import PageInfo from "../site-metadata/PageInfo";
 
 // eslint-disable-next-line react/display-name
 export default () => (
